legion/robot/requirements.txt
six>=1.12.0
urllib3>=1.24.3
requests>=2.22.0
kubernetes==10.0.1
jinja2>=2.10.1
boto3>=1.9.75
botocore>=1.12.75
robotframework>=3.0.2
robotframework-pabot>=0.44
pyyaml>=3.1.2
google-cloud-storage>=1.18.0
azure-storage-blob>=1.5.0
azure-cli-core>=2.0.75